Split bear backend into two files
JIRA: RAIL-1896
